source/requirements.txt
httpx==0.24.1
python-telegram-bot==20.3
requests==2.31.0
urllib3==2.0.2
